# backend/donald/rest/openapi.yaml
...
openapi: 3.1.0
info:
    title: Donald
    version: 1.0.0
servers:
    - url: http://localhost:8080/v1
paths:
    /agent-edges:
        get:
            tags:
                - AgentEdge
            summary: List agent-edges
            operationId: listAgentEdge
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of agent-edges
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEdgeList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - AgentEdge
            summary: Create AgentEdge
            operationId: createAgentEdge
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentEdge'
                required: true
            responses:
                "201":
                    description: Created AgentEdge
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEdge'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
    /agent-edges/{id}:
        get:
            tags:
                - AgentEdge
            summary: Get AgentEdge by ID
            operationId: getAgentEdge
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEdge'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        patch:
            tags:
                - AgentEdge
            summary: Partially update AgentEdge
            operationId: patchAgentEdge
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentEdge'
                required: true
            responses:
                "200":
                    description: Successful update
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEdge'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Version conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        delete:
            tags:
                - AgentEdge
            summary: Delete AgentEdge by ID
            operationId: deleteAgentEdge
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
    /agent-events:
        get:
            tags:
                - AgentEvent
            summary: List agent-events
            operationId: listAgentEvent
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of agent-events
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEventList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - AgentEvent
            summary: Create AgentEvent
            operationId: createAgentEvent
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentEvent'
                required: true
            responses:
                "201":
                    description: Created AgentEvent
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEvent'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
    /agent-events/{id}:
        get:
            tags:
                - AgentEvent
            summary: Get AgentEvent by ID
            operationId: getAgentEvent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEvent'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        patch:
            tags:
                - AgentEvent
            summary: Partially update AgentEvent
            operationId: patchAgentEvent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentEvent'
                required: true
            responses:
                "200":
                    description: Successful update
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEvent'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Version conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        delete:
            tags:
                - AgentEvent
            summary: Delete AgentEvent by ID
            operationId: deleteAgentEvent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
    /agent-nodes:
        get:
            tags:
                - AgentNode
            summary: List agent-nodes
            operationId: listAgentNode
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of agent-nodes
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentNodeList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - AgentNode
            summary: Create AgentNode
            operationId: createAgentNode
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentNode'
                required: true
            responses:
                "201":
                    description: Created AgentNode
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentNode'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
    /agent-nodes/{id}:
        get:
            tags:
                - AgentNode
            summary: Get AgentNode by ID
            operationId: getAgentNode
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentNode'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        patch:
            tags:
                - AgentNode
            summary: Partially update AgentNode
            operationId: patchAgentNode
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentNode'
                required: true
            responses:
                "200":
                    description: Successful update
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentNode'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Version conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        delete:
            tags:
                - AgentNode
            summary: Delete AgentNode by ID
            operationId: deleteAgentNode
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
    /agent-runs:
        get:
            tags:
                - AgentRun
            summary: List agent-runs
            operationId: listAgentRun
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of agent-runs
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentRunList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - AgentRun
            summary: Create AgentRun
            operationId: createAgentRun
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentRun'
                required: true
            responses:
                "201":
                    description: Created AgentRun
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentRun'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /artifacts:
        get:
            tags:
                - Artifact
            summary: List artifacts
            operationId: listArtifact
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of artifacts
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ArtifactList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - Artifact
            summary: Create Artifact
            operationId: createArtifact
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Artifact'
                required: true
            responses:
                "201":
                    description: Created Artifact
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Artifact'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
components:
    schemas:
        AgentEdge:
            properties:
                condition_label:
                    nullable: true
                    type: string
                created_at:
                    format: date-time
                    type: string
                created_by:
                    format: uuid
                    type: string
                edge_type:
                    $ref: '#/components/schemas/AgentEdgeType'
                from_node_uuid:
                    format: uuid
                    type: string
                run_uuid:
                    format: uuid
                    type: string
                status:
                    $ref: '#/components/schemas/AgentEdgeStatus'
                to_node_uuid:
                    format: uuid
                    type: string
                updated_at:
                    format: date-time
                    type: string
                updated_by:
                    format: uuid
                    type: string
                uuid:
                    format: uuid
                    type: string
            required:
                - uuid
                - run_uuid
                - from_node_uuid
                - to_node_uuid
                - edge_type
                - status
                - created_by
                - updated_by
                - created_at
                - updated_at
            type: object
        AgentEdgeList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/AgentEdge'
                    type: array
                next_page_token:
                    type: string
            type: object
...
        AgentEventList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/AgentEvent'
                    type: array
                next_page_token:
                    type: string
            type: object
...
        AgentNodeList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/AgentNode'
                    type: array
                next_page_token:
                    type: string
            type: object
...
        AgentRunList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/AgentRun'
                    type: array
                next_page_token:
                    type: string
            type: object
...
        ArtifactList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/Artifact'
                    type: array
                next_page_token:
                    type: string
            type: object
...
        Problem:
            properties:
                detail:
                    type: string
                instance:
                    type: string
                status:
                    type: integer
                title:
                    type: string
                type:
                    type: string
            type: object