# backend/donald/rest/server/router.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	_ "embed"

	"net/http"

	"github.com/go-chi/chi/v5"
)

//go:embed openapi.yaml
var openapiYAML []byte

func mountRoutes(r chi.Router, h *handlers) {
	r.Route("/v1", func(r chi.Router) {

		r.Get("/openapi.yaml", func(w http.ResponseWriter, r *http.Request) {
			w.Header().Set("Content-Type", "text/yaml")
			_, _ = w.Write(openapiYAML)
		})

		r.Route("/clients", func(r chi.Router) {
			r.Get("/", h.ListClient)
			r.Post("/", h.CreateClient)

			r.Route("/{id}", func(r chi.Router) {
				r.Get("/", h.GetClient)
				r.Patch("/", h.UpdateClient)
				r.Delete("/", h.DeleteClient)
			})

		})

		r.Route("/agent-runs", func(r chi.Router) {
			r.Get("/", h.ListAgentRun)
			r.Post("/", h.CreateAgentRun)

			r.Route("/{id}", func(r chi.Router) {
				r.Get("/", h.GetAgentRun)
				r.Patch("/", h.UpdateAgentRun)
				r.Delete("/", h.DeleteAgentRun)
			})

		})

		r.Route("/agent-nodes", func(r chi.Router) {
			r.Get("/", h.ListAgentNode)
			r.Post("/", h.CreateAgentNode)

			r.Route("/{id}", func(r chi.Router) {
				r.Get("/", h.GetAgentNode)
				r.Patch("/", h.UpdateAgentNode)
				r.Delete("/", h.DeleteAgentNode)
			})

		})

		r.Route("/agent-edges", func(r chi.Router) {
			r.Get("/", h.ListAgentEdge)
			r.Post("/", h.CreateAgentEdge)

			r.Route("/{id}", func(r chi.Router) {
				r.Get("/", h.GetAgentEdge)
				r.Patch("/", h.UpdateAgentEdge)
				r.Delete("/", h.DeleteAgentEdge)
			})

		})

		r.Route("/agent-events", func(r chi.Router) {
			r.Get("/", h.ListAgentEvent)
			r.Post("/", h.CreateAgentEvent)

			r.Route("/{id}", func(r chi.Router) {
				r.Get("/", h.GetAgentEvent)
				r.Patch("/", h.UpdateAgentEvent)
				r.Delete("/", h.DeleteAgentEvent)
			})

		})

		r.Route("/interventions", func(r chi.Router) {
			r.Get("/", h.ListIntervention)
			r.Post("/", h.CreateIntervention)

			r.Route("/{id}", func(r chi.Router) {
				r.Get("/", h.GetIntervention)
				r.Patch("/", h.UpdateIntervention)
				r.Delete("/", h.DeleteIntervention)
			})

		})

		r.Route("/artifacts", func(r chi.Router) {
			r.Get("/", h.ListArtifact)
			r.Post("/", h.CreateArtifact)

			r.Route("/{id}", func(r chi.Router) {
				r.Get("/", h.GetArtifact)
				r.Patch("/", h.UpdateArtifact)
				r.Delete("/", h.DeleteArtifact)
			})

		})

	})

	r.Get("/healthz", func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusOK)
		_, _ = w.Write([]byte(`{"status":"OK"}`))
	})

	// The storage package registers /upload and /sign on the default mux in
	// storage.Register. Forward those paths so they are served on the same port
	// as the REST API.
	r.Handle("/upload", http.DefaultServeMux)
	r.Handle("/sign", http.DefaultServeMux)

	// Serve the info page at the root by forwarding to the default mux, where
	// info.Register mounted it.
	r.Handle("/", http.DefaultServeMux)

}